backend-flask/server/requirements.txt
attrs
Automat==24.8.1
blinker==1.8.2
certifi==2024.8.30
cffi==1.15.1
charset-normalizer==3.3.2
click==8.1.7
constantly==23.10.4
cryptography==43.0.1
cssselect==1.2.0
defusedxml==0.7.1
filelock==3.16.1
Flask==3.0.3
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
itemadapter==0.9.0
itemloaders==1.3.1
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
lxml==5.3.0
MarkupSafe==2.1.5
packaging==24.1
parsel==1.9.1
Protego==0.3.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyDispatcher==2.0.7
pyOpenSSL==24.2.1
queuelib==1.7.0
requests==2.32.3
requests-file==2.1.0
Scrapy==2.11.2
scrapy-splash==0.9.0
service-identity==24.1.0
setuptools==75.1.0
six==1.16.0
tldextract==5.1.2
Twisted==24.7.0
typing_extensions==4.12.2
urllib3==2.2.3
w3lib==2.2.1
Werkzeug==3.0.4
zope.interface==7.0.3
pika==1.3.0
pyjwt==2.8.0
python-dotenv==1.0.0
PyJWT
